pyproject.toml
[tool.poetry]
name = "lesson-planner"
version = "0.1.0"
description = ""
authors = ["Alex Fricker"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.4"
ipython = "^8.14.0"
black = "^23.7.0"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.7"
django-cors-headers = "^4.2.0"
uwsgi = "^2.0.22"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
